+++ proxyclient/experiments/chickens.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+import sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 
-from setup import *
+from m1n1.setup import *
 
 sys_regs = dict([
     ("HID0", (3, 0, 15, 0, 0)),
