+++ proxyclient/experiments/i2c.py
@@ -1,6 +1,12 @@
-from setup import *
-import asm
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+import sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
+
 import struct
+
+from m1n1.setup import *
+from m1n1 import asm
 
 base = 0x235010000
