+++ proxyclient/experiments/memdump.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+import sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 
-import sys
-from setup import *
+from m1n1.setup import *
 
 p = 0x800000000
 limit = u.base
