+++ proxyclient/experiments/timer_test.py
@@ -1,4 +1,9 @@
-from setup import *
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+import sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
+
+from m1n1.setup import *
 
 HV_VTMR_CTL = (3, 5, 15, 1, 3)
 HV_VTMR_CTL_VMASK = (1 << 0)
