+++ proxyclient/m1n1/hv.py
@@ -1,19 +1,18 @@
-#!/usr/bin/env python3
 # SPDX-License-Identifier: MIT
-
 import sys, traceback, struct, array, bisect, os, signal
-
 from construct import *
+from enum import IntEnum, IntFlag
 
-from asm import ARMAsm
-from tgtypes import *
-from proxy import IODEV, START, EVENT, EXC, EXC_RET, ExcInfo
-from utils import *
-from sysreg import *
-from macho import MachO
-from adt import load_adt
-import xnutools
-import shell
+from .asm import ARMAsm
+from .tgtypes import *
+from .proxy import IODEV, START, EVENT, EXC, EXC_RET, ExcInfo
+from .utils import *
+from .sysreg import *
+from .macho import MachO
+from .adt import load_adt
+from . import xnutools, shell
+
+__all__ = ["HV"]
 
 PAC_MASK = 0xfffff00000000000
