+++ proxyclient/m1n1/proxyutils.py
@@ -1,11 +1,16 @@
+# SPDX-License-Identifier: MIT
 import serial, os, struct, sys, time, json, os.path, gzip, functools
-from asm import ARMAsm
-from proxy import *
-from tgtypes import *
-from sysreg import *
-from malloc import Heap
-import adt
 from contextlib import contextmanager
+from construct import *
+
+from .asm import ARMAsm
+from .proxy import *
+from .tgtypes import *
+from .sysreg import *
+from .malloc import Heap
+from . import adt
+
+__all__ = ["ProxyUtils", "RegMonitor", "GuardedHeap", "bootstrap_port"]
 
 SIMD_B = Array(32, Array(16, Int8ul))
 SIMD_H = Array(32, Array(8, Int16ul))
