--- proxyclient/setup.py
@@ -1,19 +1,0 @@
-import os, struct, sys, time
-from proxy import *
-from tgtypes import *
-from proxyutils import *
-from utils import *
-from hv import HV
-
-iface = UartInterface()
-p = M1N1Proxy(iface, debug=False)
-bootstrap_port(iface, p)
-
-u = ProxyUtils(p)
-mon = RegMonitor(u)
-hv = HV(iface, p, u)
-
-fb = u.ba.video.base
-
-print("Base at: 0x%x" % u.base)
-print("FB at: 0x%x" % fb)
