+++ proxyclient/tools/shell.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+import sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
+
+from m1n1.setup import *
+from m1n1.shell import run_shell
+
+run_shell(globals(), msg="Have fun!")
